refactor : change chrome driver logic, change vehicle matching logic
Change chrome driver logic so that if missing it'll auto install, meaning less errors for users. Now it also doesn't matter how vehicle are entered in the vehicles.py in terms of case, you could Ensure FiRe EnGines, and it'll still work. This is to make it more- easier for users. Also fixed an issue causing a mission to try to despatch even when complete. Added a Keyboard Interrupt to the bot too, so it can handle a user stopping the bot.