fix: Closest units now despatched first
By looping the checkboxes first, we now ensure the closest vehicles are despatched, rather than looping our vehicles first. So, with this new way the closest vehicle are now despatched.

Also removed the xpath check for a much more convinent check on the value of the checkbox against the vehicle id.

Added a bit of a tutorial to the config.ini for urls.