Changes to requirements to include generated missions
Generated missions are also now considered, if you have generated missions the code will take the requirements from the loaded missions. If the mission doesn't exist in the .json format, it'll manually find the requirements.

In short this is just another feature, if you use generate.bat before opening the bot essentially you shouldn't need to load requirements anymore as it'll be loaded via the json.

- Might try and add a CI / bot that'll just update the repo with missions so that perhaps users don't need to self generate.