Checking to see if mission should be despatched
We now check to see if the mission should be despatched before we even attempt it.  This will prevent us trying to do any user set missions, or any missions with vehicles currently on the way. We do this by checking the colour. Also adjusted the colour for the egg so it shows up easier.

Co-Authored-By: duchian <duchian@users.noreply.github.com>